+++ src/frequenz/client/microgrid/__init__.py
@@ -62,10 +62,10 @@ from ._exception import (
     UnknownError,
     UnrecognizedGrpcStatus,
 )
+from ._id import ComponentId, MicrogridId
 from ._metadata import Location, Metadata
 
 __all__ = [
-    "MicrogridApiClient",
     "ApiClientError",
     "BatteryComponentState",
     "BatteryData",
@@ -76,6 +76,7 @@ __all__ = [
     "Component",
     "ComponentCategory",
     "ComponentData",
+    "ComponentId",
     "ComponentMetadata",
     "ComponentMetricId",
     "ComponentType",
@@ -100,6 +101,8 @@ __all__ = [
     "Location",
     "Metadata",
     "MeterData",
+    "MicrogridApiClient",
+    "MicrogridId",
     "OperationAborted",
     "OperationCancelled",
     "OperationNotImplemented",