+++ src/frequenz/client/microgrid/_connection.py
@@ -6,15 +6,17 @@
 
 from dataclasses import dataclass
 
+from ._id import ComponentId
+
 
 @dataclass(frozen=True)
 class Connection:
     """Metadata for a connection between microgrid components."""
 
-    start: int
+    start: ComponentId
     """The component ID that represents the start component of the connection."""
 
-    end: int
+    end: ComponentId
     """The component ID that represents the end component of the connection."""
 
     def is_valid(self) -> bool:
@@ -24,4 +26,4 @@ class Connection:
             `True` if `start >= 0`, `end > 0`, and `start != end`, `False`
                 otherwise.
         """
-        return self.start >= 0 and self.end > 0 and self.start != self.end
+        return int(self.start) >= 0 and int(self.end) > 0 and self.start != self.end